Prefetch postings data.
This uses the `IndexInput#prefetch` API for postings. This relies on
heuristics, as we don't know ahead of time what data we will need from a
postings list:
 - Postings lists are prefetched entirely when they are short (< 16kB).
 - Otherwise, the first page of postings and the first page of skip data is
   prefetched.

Positions, offsets and payloads are never prefetched.

Putting the `IndexInput#prefetch` call in `TermsEnum#postings` and
`TermsEnum#impacts` works well because `BooleanQuery` will first create
postings/impacts lists for all clauses before it starts unioning/intersecting
them. This allows the prefetching logic to run in parallel across all clauses
of the same query on the same segment.