TaskExecutor should not fork for one of its tasks
When executing N tasks and waiting on the result of all of them,
we should not fork one of them and just execute at least one of them outright.
This saves at least one context switch, removes the need for any
reentrancy protection, and makes full use of the supplied task executor
if tasks fan out again.