Stop bounding outer window.
Currently `MaxScoreBulkScorer` requires its "outer" window to be at least
`WINDOW_SIZE`. The intuition there was that we should make sure we should use
the whole range of the bit set that we are using to collect matches. The
downside is that it may force us to use an upper level in the skip list that
has worse upper bounds for the scores.

luceneutil suggests that this is not a good trade-off: removing this
requirement makes some queries a bit slower, but `OrHighMin` and `OrHighRare`
much faster.