Introduce IndexSearcher#searchLeaf(LeafReaderContext, Weight, Collector) method
There's a couple of places in the codebase where we extend IndexSearcher to customize
per leaf behaviour, and in order to do that, we need to override the entire search method
that loops through the leaves. A good example is ScorerIndexSearcher.

Adding a searchLeaf method that provides the per leaf behaviour makes those cases a little
easier to deal with. Also, it paves the road for #13542 where we will want to replace
search(List<LeafReaderContext>, Weight, Collector) with something along the lines of
search(LeafReaderContextPartition[], Weight, Collector).