[9.x] Aggregate files from the same segment into a single Arena (#13682)
Backport of #13570

This commit adds a ref counted shared arena to support aggregating segment files into a single Arena.

The backport only supports the JDK 21 version of the code - JDK 19 and 20 specific versions do not aggregate files so therefore remain the same. The MemorySession/SegmentScope/Arena APIs are different in 19 and 20.