Remove IndexSearcher#search(List<LeafReaderContext>, Weight, Collector)
With the introduction of intra-segment concurrency, we have introduced a new
protected search(LeafReaderContextPartition, Weight, Collector) method. The
previous variant that accepts a list of leaf reader contexts was left deprecated
as there is one leftover usages coming from search(Query, Collector). The hope was
that the latter was going to be removed soon as well, but there is actually no
need to tie the two removals. It is easier to fold this method into its only
caller, in order for it to still bypass the collector manager based methods.