Deprecate BulkScorer#score(LeafReaderContext, Bits)
We have removed BulkScorer#score(LeafReaderContext, Bits) in main in favour of
BulkScorer#score(LeafCollector collector, Bits acceptDocs, int min, int max) as
part of #13542. This commit deprecates the method in 9.x. Internal usages of it
are left untouched as there may be subclasses that override it, which we don't
want to break in a minor release.