Add BytesRefIterator to TermInSetQuery
Addresses #13804

TermInSetQuery used to have an accessor to its terms that was removed in #12173
to protect leaking internal encoding details. This introduces an accessor to the
term data in the query that doesn't expose internal but merely allows iterating
over the decoded BytesRef, making inspection of the querys content possible again.