Remove synchronization from IndexWriter.isClosed
Blocking on this one seems unnecessary as the closed flag is volatile.
This tends to block for an extended period in its only production use
case `StandardDirectoryWriter.isCurrent()` every now and then
which seems unnecessary. Either it's true, in which case it'll stay
true and there's no need to wait for anything or it's false in which
case the else branch goes into `IndexWriter.nrtIsCurrent` which blocks
on the same mutex anyway.