Make DirectMonotonicReader.Meta more compact
Even with the all-zero bit-length optimization recently introduced, these
`Meta` instances tend to consume a non-trivial amount of heap still.
We can do away with almost 40% of their size by removing the offsets
array that can be calculated on the fly when creating the readers which
translates into tens of MB of heap or more for some ES use cases with
high segment + field count.